+++ Robust.Shared/GameObjects/Components/Appearance/AppearanceComponent.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using Robust.Shared.GameStates;
-using Robust.Shared.IoC;
 using Robust.Shared.ViewVariables;
 
 namespace Robust.Shared.GameObjects;
@@ -32,14 +31,6 @@ public sealed partial class AppearanceComponent : Component
     [ViewVariables] internal bool UpdateQueued;
 
     [ViewVariables] internal Dictionary<Enum, object> AppearanceData = new();
-
-    [Dependency] private readonly IEntitySystemManager _sysMan = default!;
-
-    [Obsolete("Use SharedAppearanceSystem instead")]
-    public void SetData(Enum key, object value)
-    {
-        _sysMan.GetEntitySystem<SharedAppearanceSystem>().SetData(Owner, key, value, this);
-    }
 
     [Obsolete("Use SharedAppearanceSystem instead")]
     public bool TryGetData<T>(Enum key, [NotNullWhen(true)] out T data)