+++ Robust.Shared/Physics/Systems/FixtureSystem.cs
@@ -5,7 +5,6 @@ using Robust.Shared.Collections;
 using Robust.Shared.GameObjects;
 using Robust.Shared.GameStates;
 using Robust.Shared.IoC;
-using Robust.Shared.Log;
 using Robust.Shared.Physics.Collision.Shapes;
 using Robust.Shared.Physics.Components;
 using Robust.Shared.Physics.Dynamics;
@@ -354,7 +353,7 @@ namespace Robust.Shared.Physics.Systems
 
             if (oldLayer != layer)
             {
-                var ev = new CollisionLayerChangeEvent(body);
+                var ev = new CollisionLayerChangeEvent((uid, body));
                 RaiseLocalEvent(ref ev);
             }