+++ Robust.UnitTesting/Shared/Physics/BroadphaseNetworkingTest.cs
@@ -65,9 +65,9 @@ public sealed class BroadphaseNetworkingTest : RobustIntegrationTest
         {
             var mapId = mapMan.CreateMap();
             map1 = mapMan.GetMapEntityId(mapId);
-            var gridComp = mapMan.CreateGrid(mapId);
-            gridComp.SetTile(Vector2i.Zero, new Tile(1));
-            grid1 = gridComp.Owner;
+            var gridEnt = mapMan.CreateGridEntity(mapId);
+            gridEnt.Comp.SetTile(Vector2i.Zero, new Tile(1));
+            grid1 = gridEnt.Owner;
         });
 
         var map1Net = sEntMan.GetNetEntity(map1);
@@ -137,9 +137,9 @@ public sealed class BroadphaseNetworkingTest : RobustIntegrationTest
             // Create grid
             var mapId = mapMan.CreateMap();
             map2 = mapMan.GetMapEntityId(mapId);
-            var gridComp = mapMan.CreateGrid(mapId);
-            gridComp.SetTile(Vector2i.Zero, new Tile(1));
-            grid2 = gridComp.Owner;
+            var gridEnt = mapMan.CreateGridEntity(mapId);
+            gridEnt.Comp.SetTile(Vector2i.Zero, new Tile(1));
+            grid2 = gridEnt.Owner;
 
             // Move player
             var coords = new EntityCoordinates(grid2, Vector2.Zero);